+++ change_log.md
@@ -1,3 +1,53 @@
+## To do:
+* message ahmed if we can submit our code tomorrow
+* push my changes
+* compare with sneaky
+* push to master
+* Jakob implement general case
+* tomorrow debug and submit
+
+* remove from bidHistory if wasn't accepted
+
+* sleep when buyer address other agent
+
+* Respond when buyer not addressing anyone
+	* Wait 2 sec
+	* Check for other agent's offer in bidHistory
+	* Exit if new offer, continue if no offer
+* Test rejectoffer
+
+## Considerations:
+* Response to different messages:
+	* General
+	* No quantity
+* Reject and increase price if buyer offer is too low
+* Track what the user bought to make better deals with? 
+* If other agent makes a SellOffer
+	* Counter it or
+	* The lowest I can go is $ because ... 
+	* Make a different deal
+
+##10/26/2020
+* When reducing price, always reduce markupRatio by 10% to 25%
+* Add price details to MinOffer in interpretMessage
+* Address human when other agent talks
+* Recognize confirmations then clear bidHistory
+
+## 10/25/2020
+* Note: Formatting for bidHistory, interpretation, offer
+``` bidHistory: {'Human': [{'quantity': {'milk': 1}, 'type': 'BuyRequest', 'metadata': {'speaker': 'Human', 'addressee': 'Watson', 'text': 'Watson I want 1 cup of milk', 'role': 'buyer', 'environmentUUID': 'abcdefg', 'timestamp': 1603411644692, 'timeStamp': 1603411646.5213463}}]},
+	interpretation: {'type': 'RejectOffer', 'metadata': {'speaker': 'Human', 'addressee': 'Watson', 'text': "Watson I don't like the offer", 'role': 'buyer', 'environmentUUID': 'abcdefg', 'timestamp': 1603411677110, 'timeStamp': 1603411677.8419251}}
+	offer: {'quantity': {'milk': 1}, 'price': {'value': 0.48, 'unit': 'USD'}, 'type': 'SellOffer', 'metadata': {'text': 'How about if I sell you 1 milk for 0.48 USD.', 'speaker': 'Watson', 'role': 'seller', 'addressee': 'Human', 'environmentUUID': 'abcdefg', 'timestamp': 1603411677799.0557, 'timeStamp': 1603411678.442418}}
+```
+* Fixed naming for myRecentOffer and myLastPrice
+* Chance of trying to convince buyer to accept price if buyer rejects it
+
+## 10/22/2020
+* If other agent makes an offer:
+	* Offer a lower price
+	* Match their offer if it's lower than our minMarkupRatio
+	* Taunt that their products are bad if they're losing money with their offer
+
 ## 10/21/2020
 ### Changes
 * Added print statements to the beginning and end of every function
@@ -12,6 +62,3 @@
 * When reach a minimum markup, return a "minOfferMessage"
 * Buyer can still accept offer after seeing the "minOfferMessage"
 
-### To do:
-* Respond when buyer not addressing anyone
-* Respond when buyer is addressing other agent